requirements.txt
kafka-python>=2.0.2
pandas>=1.5.3
scikit-learn>=1.2.0
imbalanced-learn>=0.10.0
tensorflow>=2.9.1
mlflow>=2.3.0
tensorflow-model-analysis>=0.39.0
flask>=2.2.2
prometheus_client>=0.15.0
gunicorn>=20.0.0
apache-airflow>=2.0.0
kerastuner>=1.0.0
tensorflow_privacy>=0.8.6